requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
crispy-bootstrap4==2023.1
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.4
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==23.3
django-js-asset==2.1.0
django-phonenumber-field==7.2.0
django-resized==1.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
executing==1.2.0
idna==3.4
inflection==0.5.1
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.23
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.6
